HW14 Внедрение docker и consul

+++ social-backend/src/main/java/ru/apolyakov/social_network/controller/UserController.java
@@ -58,7 +58,7 @@ public class UserController {
         if (userId == null) {
             return "redirect:/users";
         }
-        Integer currentUserId = userService.getCurrentUserId();
+        Long currentUserId = userService.getCurrentUserId();
         if (currentUserId == null || currentUserId.intValue() == userId) {
             return "redirect:/index";
         }
